Issue 6849: LTS - Remove deprecated RollingStorage implementations from bindings project. (#6858)
Remove deprecated RollingStorage implementations from bindings project.
