+++ segmentstore/server/host/src/test/java/io/pravega/segmentstore/server/host/AzureIntegrationTest.java
@@ -107,22 +107,9 @@ public class AzureIntegrationTest extends BookKeeperIntegrationTestBase {
      * AzureChunkStorage does not support rolling storage.
      */
     @Override
-    public void testEndToEnd() {
-    }
-
-    /**
-     * AzureChunkStorage does not support rolling storage.
-     */
-    @Override
     public void testFlushToStorage() {
     }
 
-    /**
-     * AzureChunkStorage does not support rolling storage.
-     */
-    @Override
-    public void testEndToEndWithFencing() {
-    }
     //endregion
 
     private class LocalAzureSimpleStorageFactory implements SimpleStorageFactory {
