+++ segmentstore/server/host/src/test/java/io/pravega/segmentstore/server/host/FileSystemIntegrationTest.java
@@ -22,10 +22,8 @@ import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperConfig;
 import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperLogFactory;
 import io.pravega.storage.filesystem.FileSystemSimpleStorageFactory;
 import io.pravega.storage.filesystem.FileSystemStorageConfig;
-import io.pravega.storage.filesystem.FileSystemStorageFactory;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Test;
 
 import java.time.Duration;
 
@@ -57,8 +55,7 @@ public class FileSystemIntegrationTest extends BookKeeperIntegrationTestBase {
 
         return ServiceBuilder
                 .newInMemoryBuilder(builderConfig)
-                .withStorageFactory(setup -> useChunkedSegmentStorage ?
-                        new FileSystemSimpleStorageFactory(ChunkedSegmentStorageConfig.DEFAULT_CONFIG.toBuilder()
+                .withStorageFactory(setup -> new FileSystemSimpleStorageFactory(ChunkedSegmentStorageConfig.DEFAULT_CONFIG.toBuilder()
                                 .journalSnapshotInfoUpdateFrequency(Duration.ofMillis(10))
                                 .maxJournalUpdatesPerSnapshot(5)
                                 .garbageCollectionDelay(Duration.ofMillis(10))
@@ -66,21 +63,8 @@ public class FileSystemIntegrationTest extends BookKeeperIntegrationTestBase {
                                 .selfCheckEnabled(true)
                                 .build(),
                                 setup.getConfig(FileSystemStorageConfig::builder),
-                                setup.getStorageExecutor())
-                        : new FileSystemStorageFactory(setup.getConfig(FileSystemStorageConfig::builder), setup.getStorageExecutor())
-                )
+                                setup.getStorageExecutor()))
                 .withDataLogFactory(setup -> new BookKeeperLogFactory(setup.getConfig(BookKeeperConfig::builder),
                         getBookkeeper().getZkClient(), setup.getCoreExecutor()));
-    }
-
-    /**
-     * SegmentStore is used to create some segments, write data to them and let them flush to the storage.
-     * This test only uses this storage to restore the container metadata segments in a new durable data log. Segment
-     * properties are matched for verification after the restoration.
-     * @throws Exception If an exception occurred.
-     */
-    @Test
-    public void testDataRecovery() throws Exception {
-        testSegmentRestoration();
     }
 }
