+++ segmentstore/server/host/src/test/java/io/pravega/segmentstore/server/host/HDFSIntegrationTest.java
@@ -23,12 +23,10 @@ import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperLogFactory;
 import io.pravega.storage.hdfs.HDFSClusterHelpers;
 import io.pravega.storage.hdfs.HDFSSimpleStorageFactory;
 import io.pravega.storage.hdfs.HDFSStorageConfig;
-import io.pravega.storage.hdfs.HDFSStorageFactory;
 import lombok.val;
 import org.apache.hadoop.hdfs.MiniDFSCluster;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Test;
 
 import java.time.Duration;
 
@@ -83,8 +81,7 @@ public class HDFSIntegrationTest extends BookKeeperIntegrationTestBase {
         ServiceBuilderConfig builderConfig = getBuilderConfig(configBuilder, instanceId);
         return ServiceBuilder
                 .newInMemoryBuilder(builderConfig)
-                .withStorageFactory(setup -> useChunkedSegmentStorage ?
-                        new HDFSSimpleStorageFactory(ChunkedSegmentStorageConfig.DEFAULT_CONFIG.toBuilder()
+                .withStorageFactory(setup -> new HDFSSimpleStorageFactory(ChunkedSegmentStorageConfig.DEFAULT_CONFIG.toBuilder()
                                 .journalSnapshotInfoUpdateFrequency(Duration.ofMillis(10))
                                 .maxJournalUpdatesPerSnapshot(5)
                                 .garbageCollectionDelay(Duration.ofMillis(10))
@@ -92,20 +89,8 @@ public class HDFSIntegrationTest extends BookKeeperIntegrationTestBase {
                                 .selfCheckEnabled(true)
                                 .build(),
                                 setup.getConfig(HDFSStorageConfig::builder),
-                                setup.getStorageExecutor())
-                        : new HDFSStorageFactory(setup.getConfig(HDFSStorageConfig::builder), setup.getStorageExecutor()))
+                                setup.getStorageExecutor()))
                 .withDataLogFactory(setup -> new BookKeeperLogFactory(setup.getConfig(BookKeeperConfig::builder), getBookkeeper().getZkClient(), setup.getCoreExecutor()));
-    }
-
-    /**
-     * SegmentStore is used to create some segments, write data to them and let them flush to the storage.
-     * This test only uses this storage to restore the container metadata segments in a new durable data log. Segment
-     * properties are matched for verification after the restoration.
-     * @throws Exception If an exception occurred.
-     */
-    @Test(timeout = 120000)
-    public void testDataRecovery() throws Exception {
-        testSegmentRestoration();
     }
     //endregion
 }
